requirements_2.txt
langchain
langchain_community
langchain-groq
streamlit
faiss-cpu
langchain_cohere
bs4
google.generativeai